--- a/openquake/hazardlib/source/point.py
+++ b/openquake/hazardlib/source/point.py
@@ -252,7 +252,8 @@ class PointSource(ParametricSeismicSource):
                             0, np.rake, occurrence_rate,
                             self.temporal_occurrence_model)
                     else:
-                        surface, nhc = self._get_rupture_surface(mag, np, hc)
+                        surface, nhc = self._get_rupture_surface(
+                            mag, np, hc, kwargs.get('shift_hypo'))
                         yield ParametricProbabilisticRupture(
                             mag, np.rake, self.tectonic_region_type,
                             nhc if kwargs.get('shift_hypo') else hc,
@@ -286,7 +287,8 @@ class PointSource(ParametricSeismicSource):
         """
         return len(self.get_annual_occurrence_rates()) * self.count_nphc()
 
-    def _get_rupture_surface(self, mag, nodal_plane, hypocenter):
+    def _get_rupture_surface(
+            self, mag, nodal_plane, hypocenter, shift_hypo=False):
         """
         Create and return rupture surface object with given properties.
 
@@ -298,8 +300,10 @@ class PointSource(ParametricSeismicSource):
             describing the rupture orientation.
         :param hypocenter:
             Point representing rupture's hypocenter.
+        :param shift_hypo:
+            If true, returns the shifted hypocenter
         :returns:
-            Instance of :class:`~openquake.hazardlib.geo.surface.planar.PlanarSurface`.
+            Tuple (PlanarSurface, hypocenter)
         """
         eps = .001  # 1 meter buffer to survive numerical errors
         assert self.upper_seismogenic_depth < hypocenter.depth + eps, (
@@ -319,6 +323,8 @@ class PointSource(ParametricSeismicSource):
         rup_length, rup_proj_width, rup_proj_height = _get_rupture_dimensions(
             self, mag, nodal_plane.rake, nodal_plane.dip)
 
+        rupture_center = hypocenter
+
         # half height of the vertical component of rupture width
         # is the vertical distance between the rupture geometrical
         # center and it's upper and lower borders:
@@ -331,18 +337,18 @@ class PointSource(ParametricSeismicSource):
         if vshift < 0:
             # the top edge is below upper seismogenic depth. now we need
             # to check that we do not cross the lower border.
-            vshift = self.lower_seismogenic_depth - hypocenter.depth - hheight
+            vshift = (self.lower_seismogenic_depth -
+                      hypocenter.depth - hheight)
             if vshift > 0:
-                # the bottom edge of the rupture is above the lower sesmogenic
+                # the bottom edge of the rupture is above the lower seismo
                 # depth. that means that we don't need to move the rupture
                 # as it fits inside seismogenic layer.
                 vshift = 0
-            # if vshift < 0 than we need to move the rupture up by that value.
+            # if vshift < 0 than we need to move the rupture up.
 
         # now we need to find the position of rupture's geometrical center.
         # in any case the hypocenter point must lie on the surface, however
         # the rupture center might be off (below or above) along the dip.
-        rupture_center = hypocenter
         if vshift != 0:
             # we need to move the rupture center to make the rupture fit
             # inside the seismogenic layer.
@@ -383,8 +389,8 @@ class PointSource(ParametricSeismicSource):
             azimuth=(nodal_plane.strike + theta) % 360)
         surface = PlanarSurface(
             nodal_plane.strike, nodal_plane.dip, left_top, right_top,
-            right_bottom, left_bottom)
-        return surface, rupture_center
+            right_bottom, left_bottom, check=False)
+        return surface, rupture_center if shift_hypo else hypocenter
 
     @property
     def polygon(self):