Part of https://github.com/gem/oq-engine/issues/7745. This avoids redundant calls to iruptures and improves a bit calculations with no nodal plane/hypocenter uncertanties (when there is no need to call `_cps_rups`):
```
$ OQ_SAMPLE_SITES=.001 oq run GLD.zip -c0 && oq show performance
# master
| calc_873, maxmem=0.5 GB    | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| ClassicalCalculator.run    | 67.8     | 245.8     | 1      |
| total classical            | 64.8     | 210.4     | 1      |
| make_contexts              | 41.4     | 195.9     | 2_516  |

# this branch
| calc_872, maxmem=0.4 GB    | time_sec | memory_mb | counts |
|----------------------------+----------+-----------+--------|
| ClassicalCalculator.run    | 56.4     | 149.6     | 1      |
| total classical            | 53.3     | 210.6     | 1      |
| make_contexts              | 29.3     | 194.1     | 2_516  |
```
There is also some cleanup and usage of the @ operator which gives a marginal speedup.